login socialite

+++ ecmonce/database/migrations/2017_04_15_012122_edit_price_to_products.php
@@ -26,7 +26,7 @@ class EditPriceToProducts extends Migration
     public function down()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('price')->change();
+            $table->decimal('price', 5, 2)->change();
         });
     }
 }
